Fix threading bug.
The cause was that the client thread sometimes went out of scope before
one of its service workers which would lead to a mutex being unlocked
after it was freed which lead to a core dump. The service workers are
now saved in a std::vector which is iterated through in the destructor
and each worker is joined on to make sure it has stopped.