Fix bug where definitions were written out of order.
The issue was that the set does not preserve order which resulted in
nested types being written after their enclosing types. Now, they are
added to a list which preserves order and so the inner most nested type
is examined last. Then they are written in the reverse order when
writing their corresponding LabComm definitions which gives the correct
order.